G35: Add (V)erbose parameter
As DEBUG_LEVELING creates a *lot* of output, it is helpful to add a
Verbose parameter to this command that gives more fine-grained control.

As tramming (leveling) the bed is a complex process, verbose output will
often be helpful.

This commit introduces a verbosity level of 0-4, default 0. Currently,
only 0 and >0 makes a difference to the output.